requirements.txt
requests
tqdm
python-dotenv
discord
